requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
gitdb==4.0.10
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
smmap==5.0.0
sqlparse==0.4.4
uritemplate==4.1.1
